requirements.txt
python-dotenv
kiteconnect
starlette
uvicorn
fastapi
rich
google-adk
requests
pydantic
Flask